Restructure js/libsql-client to allow importing standalone HTTP client.
