+++ packages/js/libsql-client/package-cjs.json
@@ -1,0 +1,3 @@
+{
+    "type": "commonjs"
+}
