+++ packages/js/libsql-client/tsconfig.build-esm.json
@@ -1,0 +1,9 @@
+{
+    "extends": "./tsconfig.base.json",
+    "compilerOptions": {
+        "module": "esnext",
+        "declaration": true,
+        "outDir": "./lib-esm/"
+    }
+}
+
